+++ test/devicecontroller/DefaultDeviceController.test.ts
@@ -132,11 +132,7 @@ describe('DefaultDeviceController', () => {
     });
 
     it('returns an empty list if MediaDeviceInfo API does not exist', async () => {
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const GlobalAny = global as any;
-      // Deleting MediaDeviceInfo throws "ReferenceError: MediaDeviceInfo is not defined."
-      // For testing, assign a boolean.
-      GlobalAny['window']['MediaDeviceInfo'] = false;
+      MediaDeviceInfo = undefined;
       const devices: MediaDeviceInfo[] = await deviceController.listAudioInputDevices();
       expect(devices.length).to.equal(0);
     });
@@ -577,6 +573,14 @@ describe('DefaultDeviceController', () => {
         await deviceController.chooseAudioInputDevice(device);
       } catch (e) {
         throw new Error('This line should not be reached.');
+      }
+    });
+
+    it('pass undefined device', async () => {
+      try {
+        await deviceController.chooseAudioInputDevice(undefined);
+      } catch (e) {
+        expect(e.message).to.be.eq('Undefined device');
       }
     });